+++ packages/abi-gen/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0 - _???_
 
 * Added CLI options for explicit specifying location of partials and main template (#346)
+* Added CLI option to specify networkId, adding support for the JSON artifact format found in @0xproject/contracts
 
 ## v0.1.0 - _January 11, 2018_